[cocoemu-server] Rename ServerAdapter to Cdm16ServerAdapter, change it logic
Cdm16ServerAdapter now implements processorInfo interface

Fix all related with it problems in ServerMessageHandler
